+++ changelog.md
@@ -3,6 +3,7 @@
 * Let wither change target if can't see it for x secs and not player
 
 ## Upcoming
+* Added tooltip to Dragon Egg to let you know that more drop from dragons
 * Minor update for latest InsaneLib
   * Please note that some config options have been renamed, so you have to set them again